Fix database schema and migration issues
- Consolidated database connection logic into lib/db/connection.ts
- Updated schema.ts to use ai_chatbot schema consistently
- Fixed package.json build script to use auto-migrate.ts
- Renamed migration file to match journal expectations
- Updated migration metadata to use ai_chatbot schema
- Cleaned up unused migration files and backups
- Added auto-migration script for seamless database setup
