+++ .env.example
@@ -1,18 +1,38 @@
-# The following keys below are automatically created and
-# added to your environment when you deploy on vercel
+# ============================================
+# REQUIRED: Databricks Configuration
+# ============================================
+# Your Databricks workspace hostname (without https://)
+DATABRICKS_HOST=your-workspace.cloud.databricks.com
 
+# Option 1: OAuth application credentials (recommended)
+DATABRICKS_CLIENT_ID=your-oauth-client-id
+DATABRICKS_CLIENT_SECRET=your-oauth-client-secret
+
+# Option 2: Personal Access Token (alternative to OAuth)
+# DATABRICKS_TOKEN=your-personal-access-token
+
+# ============================================
+# REQUIRED: PostgreSQL Configuration
+# ============================================
+# Option 1: Use individual PG* variables (recommended for Databricks OAuth)
+PGHOST=your-postgres-host
+PGDATABASE=your-database-name
+PGUSER=your-username
+PGPORT=5432
+# Note: PGPASSWORD is automatically set via OAuth token exchange
+
+# Option 2: Use full connection string (alternative to PG* variables)
+# POSTGRES_URL=postgresql://username:password@host:port/database
+
+# ============================================
+# OPTIONAL: Additional Services
+# ============================================
 # AI Gateway API Key (required for non-Vercel deployments)
 # For Vercel deployments, OIDC tokens are used automatically
 AI_GATEWAY_API_KEY=****
 
-
-# Instructions to create a Vercel Blob Store here: https://vercel.com/docs/storage/vercel-blob
+# Blob storage for file uploads (optional)
 BLOB_READ_WRITE_TOKEN=****
 
-# Instructions to create a PostgreSQL database here: https://vercel.com/docs/storage/vercel-postgres/quickstart
-POSTGRES_URL=****
-
-
-# Instructions to create a Redis store here:
-# https://vercel.com/docs/redis
+# Redis for resumable streams (optional)
 REDIS_URL=****
