+++ app.yaml
@@ -1,4 +1,4 @@
-command: ["sh", "-c", "npm run db:migrate && npm run start"]
+command: ["npm", "run", "start"]
 runtime: nodejs20
 
 resources:
